+++ pkg/commands/assert/command.go
@@ -11,9 +11,9 @@ import (
 	tclient "github.com/kyverno/chainsaw/pkg/client/testing"
 	engineclient "github.com/kyverno/chainsaw/pkg/engine/client"
 	nspacer "github.com/kyverno/chainsaw/pkg/engine/namespacer"
+	opassert "github.com/kyverno/chainsaw/pkg/engine/operations/assert"
 	"github.com/kyverno/chainsaw/pkg/loaders/config"
 	"github.com/kyverno/chainsaw/pkg/loaders/resource"
-	opassert "github.com/kyverno/chainsaw/pkg/runner/operations/assert"
 	restutils "github.com/kyverno/chainsaw/pkg/utils/rest"
 	"github.com/kyverno/pkg/ext/output/color"
 	"github.com/spf13/cobra"
