+++ pkg/engine/operations/assert/operation_test.go
