+++ pkg/engine/operations/create/operation.go
@@ -1,22 +1,23 @@
-package patch
+package create
 
 import (
 	"context"
+	"errors"
 
 	"github.com/jmespath-community/go-jmespath/pkg/binding"
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
+	"github.com/kyverno/chainsaw/pkg/cleanup/cleaner"
 	"github.com/kyverno/chainsaw/pkg/client"
 	apibindings "github.com/kyverno/chainsaw/pkg/engine/bindings"
 	"github.com/kyverno/chainsaw/pkg/engine/checks"
 	"github.com/kyverno/chainsaw/pkg/engine/logging"
 	"github.com/kyverno/chainsaw/pkg/engine/namespacer"
+	"github.com/kyverno/chainsaw/pkg/engine/operations"
+	"github.com/kyverno/chainsaw/pkg/engine/operations/internal"
 	"github.com/kyverno/chainsaw/pkg/engine/outputs"
 	"github.com/kyverno/chainsaw/pkg/engine/templating"
-	"github.com/kyverno/chainsaw/pkg/runner/operations"
-	"github.com/kyverno/chainsaw/pkg/runner/operations/internal"
+	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/json"
 	"k8s.io/apimachinery/pkg/util/wait"
 )
 
@@ -24,6 +25,7 @@ type operation struct {
 	client     client.Client
 	base       unstructured.Unstructured
 	namespacer namespacer.Namespacer
+	cleaner    cleaner.CleanerCollector
 	template   bool
 	expect     []v1alpha1.Expectation
 	outputs    []v1alpha1.Output
@@ -33,6 +35,7 @@ func New(
 	client client.Client,
 	obj unstructured.Unstructured,
 	namespacer namespacer.Namespacer,
+	cleaner cleaner.CleanerCollector,
 	template bool,
 	expect []v1alpha1.Expectation,
 	outputs []v1alpha1.Output,
@@ -41,6 +44,7 @@ func New(
 		client:     client,
 		base:       obj,
 		namespacer: namespacer,
+		cleaner:    cleaner,
 		template:   template,
 		expect:     expect,
 		outputs:    outputs,
@@ -54,7 +58,7 @@ func (o *operation) Exec(ctx context.Context, bindings binding.Bindings) (_ outp
 	obj := o.base
 	logger := internal.GetLogger(ctx, &obj)
 	defer func() {
-		internal.LogEnd(logger, logging.Patch, _err)
+		internal.LogEnd(logger, logging.Create, _err)
 	}()
 	if o.template {
 		template := v1alpha1.Any{
@@ -69,15 +73,15 @@ func (o *operation) Exec(ctx context.Context, bindings binding.Bindings) (_ outp
 	if err := internal.ApplyNamespacer(o.namespacer, o.client, &obj); err != nil {
 		return nil, err
 	}
-	internal.LogStart(logger, logging.Patch)
+	internal.LogStart(logger, logging.Create)
 	return o.execute(ctx, bindings, obj)
 }
 
 func (o *operation) execute(ctx context.Context, bindings binding.Bindings, obj unstructured.Unstructured) (outputs.Outputs, error) {
 	var lastErr error
 	var outputs outputs.Outputs
-	err := wait.PollUntilContextCancel(ctx, internal.PollInterval, false, func(ctx context.Context) (bool, error) {
-		outputs, lastErr = o.tryPatchResource(ctx, bindings, obj)
+	err := wait.PollUntilContextCancel(ctx, client.PollInterval, false, func(ctx context.Context) (bool, error) {
+		outputs, lastErr = o.tryCreateResource(ctx, bindings, obj)
 		// TODO: determine if the error can be retried
 		return lastErr == nil, nil
 	})
@@ -90,26 +94,26 @@ func (o *operation) execute(ctx context.Context, bindings binding.Bindings, obj 
 	return outputs, err
 }
 
-func (o *operation) tryPatchResource(ctx context.Context, bindings binding.Bindings, obj unstructured.Unstructured) (outputs.Outputs, error) {
+// TODO: could be replaced by checking the already exists error
+func (o *operation) tryCreateResource(ctx context.Context, bindings binding.Bindings, obj unstructured.Unstructured) (outputs.Outputs, error) {
 	var actual unstructured.Unstructured
 	actual.SetGroupVersionKind(obj.GetObjectKind().GroupVersionKind())
 	err := o.client.Get(ctx, client.Key(&obj), &actual)
-	if err != nil {
-		return nil, err
+	if err == nil {
+		return nil, errors.New("the resource already exists in the cluster")
 	}
-	return o.updateResource(ctx, bindings, &actual, obj)
+	if kerrors.IsNotFound(err) {
+		return o.createResource(ctx, bindings, obj)
+	}
+	return nil, err
 }
 
-func (o *operation) updateResource(ctx context.Context, bindings binding.Bindings, actual *unstructured.Unstructured, obj unstructured.Unstructured) (outputs.Outputs, error) {
-	patched, err := client.PatchObject(actual, &obj)
-	if err != nil {
-		return nil, err
+func (o *operation) createResource(ctx context.Context, bindings binding.Bindings, obj unstructured.Unstructured) (outputs.Outputs, error) {
+	err := o.client.Create(ctx, &obj)
+	if err == nil && o.cleaner != nil {
+		o.cleaner.Add(o.client, &obj)
 	}
-	bytes, err := json.Marshal(patched)
-	if err != nil {
-		return nil, err
-	}
-	return o.handleCheck(ctx, bindings, obj, o.client.Patch(ctx, actual, client.RawPatch(types.MergePatchType, bytes)))
+	return o.handleCheck(ctx, bindings, obj, err)
 }
 
 func (o *operation) handleCheck(ctx context.Context, bindings binding.Bindings, obj unstructured.Unstructured, err error) (_outputs outputs.Outputs, _err error) {
