+++ pkg/engine/operations/delete/operation_test.go
