+++ pkg/engine/operations/internal/command_output_test.go
