+++ pkg/engine/operations/script/operation_test.go
