--- pkg/runner/operations/internal/interval.go
@@ -1,7 +1,0 @@
-package internal
-
-import (
-	"time"
-)
-
-const PollInterval = 50 * time.Millisecond
