+++ pkg/runner/processors/step.go
@@ -13,20 +13,20 @@ import (
 	"github.com/kyverno/chainsaw/pkg/engine/kubectl"
 	"github.com/kyverno/chainsaw/pkg/engine/logging"
 	"github.com/kyverno/chainsaw/pkg/engine/namespacer"
+	"github.com/kyverno/chainsaw/pkg/engine/operations"
+	opapply "github.com/kyverno/chainsaw/pkg/engine/operations/apply"
+	opassert "github.com/kyverno/chainsaw/pkg/engine/operations/assert"
+	opcommand "github.com/kyverno/chainsaw/pkg/engine/operations/command"
+	opcreate "github.com/kyverno/chainsaw/pkg/engine/operations/create"
+	opdelete "github.com/kyverno/chainsaw/pkg/engine/operations/delete"
+	operror "github.com/kyverno/chainsaw/pkg/engine/operations/error"
+	oppatch "github.com/kyverno/chainsaw/pkg/engine/operations/patch"
+	opscript "github.com/kyverno/chainsaw/pkg/engine/operations/script"
+	opsleep "github.com/kyverno/chainsaw/pkg/engine/operations/sleep"
+	opupdate "github.com/kyverno/chainsaw/pkg/engine/operations/update"
 	"github.com/kyverno/chainsaw/pkg/loaders/resource"
 	"github.com/kyverno/chainsaw/pkg/report"
 	"github.com/kyverno/chainsaw/pkg/runner/failer"
-	"github.com/kyverno/chainsaw/pkg/runner/operations"
-	opapply "github.com/kyverno/chainsaw/pkg/runner/operations/apply"
-	opassert "github.com/kyverno/chainsaw/pkg/runner/operations/assert"
-	opcommand "github.com/kyverno/chainsaw/pkg/runner/operations/command"
-	opcreate "github.com/kyverno/chainsaw/pkg/runner/operations/create"
-	opdelete "github.com/kyverno/chainsaw/pkg/runner/operations/delete"
-	operror "github.com/kyverno/chainsaw/pkg/runner/operations/error"
-	oppatch "github.com/kyverno/chainsaw/pkg/runner/operations/patch"
-	opscript "github.com/kyverno/chainsaw/pkg/runner/operations/script"
-	opsleep "github.com/kyverno/chainsaw/pkg/runner/operations/sleep"
-	opupdate "github.com/kyverno/chainsaw/pkg/runner/operations/update"
 	"github.com/kyverno/chainsaw/pkg/runner/timeout"
 	"github.com/kyverno/chainsaw/pkg/testing"
 	"github.com/kyverno/pkg/ext/output/color"
